Re-enabled simulation unit tests
