+++ src/test/java/com/vmware/bespin/scheduler/TestScheduler.java
@@ -62,50 +62,6 @@ public class TestScheduler {
         assertFalse(scheduler.checkForCapacityViolation());
     }
 
-    /* TODO: move to simuation
-    @Test
-    public void testRunnerInstantiation() throws ClassNotFoundException {
-        final long NUM_NODES = 2;
-        final long CORES_PER_NODE = 3;
-        final long MEMSLICES_PER_NODE = 4;
-        final long NUM_APPS = 5;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        Scheduler scheduler = new Scheduler(conn, null);
-
-        // Check number of nodes
-        assertEquals(runner.numNodes(), NUM_NODES);
-        
-        // Check core capacity and use
-        assertEquals(runner.coreCapacity(), NUM_NODES * CORES_PER_NODE);
-        assertEquals(runner.usedCores(), 0);
-
-        // Check memslice capacity
-        assertEquals(runner.memsliceCapacity(), NUM_NODES * MEMSLICES_PER_NODE);
-        assertEquals(runner.usedMemslices(), 0);
-
-        // Check per-node valudes
-        for (long i = 1; i <= NUM_NODES; i++) {
-            assertEquals(runner.coreCapacityForNode(i), CORES_PER_NODE);
-            assertEquals(runner.usedCoresForNode(i), 0);
-            assertEquals(runner.memsliceCapacityForNode(i), MEMSLICES_PER_NODE);
-            assertEquals(runner.usedMemslicesForNode(i), 0);
-        }
-
-        // Check number of applications
-        assertEquals(runner.numApps(), NUM_APPS);
-
-        // Check pending requests
-        assertEquals(runner.getNumPendingRequests(), 0);
-        long[] idList = runner.getPendingRequestIDs();
-        assertEquals(0, idList.length);
-
-        // Check capacity
-        assertFalse(runner.checkForCapacityViolation());
-    }
-    */
-
     @Test
     public void testAddNode() throws ClassNotFoundException {
         long numNodes = 0;
@@ -226,78 +182,6 @@ public class TestScheduler {
         }
     }
 
-    /* TODO: move to simulation tests
-    @Test
-    public void testChooseRandomApplication() throws ClassNotFoundException {
-        final long NUM_NODES = 2;
-        final long CORES_PER_NODE = 3;
-        final long MEMSLICES_PER_NODE = 4;
-        final long NUM_APPS = 5;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        DiNOSSolver solver = new DiNOSSolver(conn, false, false, false);
-        Scheduler scheduler = new Scheduler(conn, solver);
-
-        // Populate applications
-        for (int i = 1; i <= NUM_APPS; i++) {
-            scheduler.addApplication(i);
-            assertEquals(scheduler.numApps(), i);
-        }
-
-        // Check number of applications
-        for (int i = 0; i < 10; i++) {
-            final long randomApp = scheduler.chooseRandomApplication();
-            assert(randomApp >= 1 && randomApp <= NUM_APPS);
-        }
-
-        // Check number of applications
-        assertEquals(runner.numApps(), NUM_APPS);
-    }
-
-    @Test
-    public void testCoresForUtil() throws ClassNotFoundException {
-        final long NUM_NODES = 2;
-        final long CORES_PER_NODE = 3;
-        final long MEMSLICES_PER_NODE = 4;
-        final long NUM_APPS = 5;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        DCMScheduler runner = new DCMScheduler(conn, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-            NUM_APPS, null, false, false);
-
-        // check cores for util at 100%
-        assertEquals(runner.coresForUtil(100), runner.coreCapacity());
-        assertEquals(runner.coresForUtil(100), NUM_NODES * CORES_PER_NODE);
-
-        // check cores for util at 50%
-        assertEquals(runner.coresForUtil(50), runner.coreCapacity() / 2);        
-        assertEquals(runner.coresForUtil(50), (NUM_NODES * CORES_PER_NODE) / 2);
-    }
-
-    @Test
-    public void testMemslicesForUtil() throws ClassNotFoundException {
-        final long NUM_NODES = 2;
-        final long CORES_PER_NODE = 3;
-        final long MEMSLICES_PER_NODE = 4;
-        final long NUM_APPS = 5;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        DCMScheduler runner = new DCMScheduler(conn, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-            NUM_APPS, null, false, false);
-
-        // check cores for util at 100%
-        assertEquals(runner.memslicesForUtil(100), runner.memsliceCapacity());
-        assertEquals(runner.memslicesForUtil(100), NUM_NODES * MEMSLICES_PER_NODE);
-
-        // check cores for util at 50%
-        assertEquals(runner.memslicesForUtil(50), runner.memsliceCapacity() / 2);        
-        assertEquals(runner.memslicesForUtil(50), (NUM_NODES * MEMSLICES_PER_NODE) / 2);
-    }
-    */
-
     @Test
     public void testPendingRequests() throws ClassNotFoundException {
         final long NUM_NODES = 2;
@@ -387,27 +271,6 @@ public class TestScheduler {
             pendingIdSetBaseline.add(i);
         }
     }
-
-    /* TODO: add to simulator tests
-    @Test
-    public void testPendingRandomRequests() throws ClassNotFoundException {
-        final long NUM_NODES = 2;
-        final long CORES_PER_NODE = 3;
-        final long MEMSLICES_PER_NODE = 4;
-        final long NUM_APPS = 5;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        DCMScheduler runner = new DCMScheduler(conn, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-            NUM_APPS, null, false, false);
-
-        // Add a request
-        for (long i = 0; i < runner.coreCapacity() + runner.memsliceCapacity(); i++) {
-            assertEquals(runner.getNumPendingRequests(), i);
-            runner.generateRandomRequest();
-        }
-    }
-    */
 
     @Test
     public void testNothingToSolve() throws Exception {
@@ -513,65 +376,6 @@ public class TestScheduler {
             assertFalse(scheduler.checkForCapacityViolation());
         }
     }
-
-    /* TODO: rewrite with simulation
-    @Test
-    public void testReleaseAllocation() throws Exception {
-        final long NUM_NODES = 5;
-        final long CORES_PER_NODE = 5;
-        final long MEMSLICES_PER_NODE = 5;
-        final long NUM_APPS = 5;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        Scheduler scheduler = new Scheduler(conn, null);
-        
-        final Random rand = new Random(0xc0ffee);
-
-        // Fill entire cluster
-        runner.fillRandom(100);
-
-        for (long application = 1; application <= NUM_APPS; application++) {
-            for (long node = 1; node <= NUM_NODES; node++) {
-                final long cores = scheduler.usedCoresForApplicationOnNode(application, node);
-                final long memslices = scheduler.usedMemslicesForApplicationOnNode(application, node);
-    
-                final long coresToRelease;
-                final long memslicesToRelease;
-                if (0 == cores) {
-                    coresToRelease = 0;
-                } else {
-                    coresToRelease = (long) rand.nextInt((int) cores + 1);
-                }
-                if (0 == memslices) {
-                    memslicesToRelease = 0;
-                } else {
-                    memslicesToRelease = (long) rand.nextInt((int) memslices + 1);
-                }
-
-                final long initialCoreCapacity = scheduler.coreCapacity();
-                final long initialMemsliceCapacity = scheduler.memsliceCapacity();
-                final long initialUsedCores = scheduler.usedCores();
-                final long initialUsedMemslices = scheduler.usedMemslices();
-                final long initialCoresForApplication = scheduler.usedCoresForApplication(application);
-                final long initialMemslicesForApplication = scheduler.usedMemslicesForApplication(application);
-                final long initialCoresForNode = scheduler.usedCoresForNode(node);
-                final long initialMemslicesForNode = scheduler.usedMemslicesForNode(node);
-                scheduler.releaseAllocation(node, application, coresToRelease, memslicesToRelease);
-                assertEquals(initialCoreCapacity, scheduler.coreCapacity());
-                assertEquals(initialMemsliceCapacity, scheduler.memsliceCapacity());
-                assertEquals(initialUsedCores - coresToRelease, scheduler.usedCores());
-                assertEquals(initialUsedMemslices - memslicesToRelease, scheduler.usedMemslices());
-                assertEquals(initialCoresForApplication - coresToRelease, scheduler.usedCoresForApplication(application));
-                assertEquals(initialMemslicesForApplication - memslicesToRelease, scheduler.usedMemslicesForApplication(application));
-                assertEquals(initialCoresForNode - coresToRelease, scheduler.usedCoresForNode(node));
-                assertEquals(initialMemslicesForNode - memslicesToRelease, scheduler.usedMemslicesForNode(node));
-                assertEquals(cores - coresToRelease, scheduler.usedCoresForApplicationOnNode(application, node));
-                assertEquals(memslices - memslicesToRelease, scheduler.usedMemslicesForApplicationOnNode(application, node));
-            }
-        }
-    }
-    */
 
     @Test
     public void testApplication() throws Exception {
@@ -807,208 +611,4 @@ public class TestScheduler {
         assertEquals(scheduler.usedMemslices(), 4);
         assertEquals(scheduler.getNumPendingRequests(), 0);
     }
-
-    /* TODO: add to simulator test
-    @Test
-    public void testFillRandomUtil() throws Exception {
-        final long NUM_NODES = 10;
-        final long CORES_PER_NODE = 10;
-        final long MEMSLICES_PER_NODE = 20;
-        final long NUM_APPS = 10;
-
-        // Create database
-        DSLContext conn = DBUtils.getConn();
-        DCMScheduler runner1 = new DCMScheduler(conn1, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-            NUM_APPS, null, false, false);
-
-        runner1.fillRandom(100);
-
-        // Make sure filled all the way.
-        assertEquals(runner1.coreCapacity(), runner1.usedCores());
-        assertEquals(runner1.memsliceCapacity(), runner1.usedMemslices());
-    }
-
-    @Test
-    public void testFillRandomDistribution() throws Exception {
-        final long NUM_NODES = 10;
-        final long CORES_PER_NODE = 10;
-        final long MEMSLICES_PER_NODE = 20;
-        final long NUM_APPS = 10;
-
-        // Test randomness of fill per node.
-        // Get the average of averages of number of cores on a node each iter.
-        int[] aggregate_core_fill = new int[(int) NUM_NODES];
-        int[] aggregate_memslice_fill = new int[(int) NUM_NODES];
-
-        int[] aggregate_core_per_app = new int[(int) NUM_APPS];
-        int[] aggregate_memslice_per_app = new int[(int) NUM_APPS];
-
-        final int ITERS = 100;
-        final int FILL_UTIL = 50;
-
-        for (long i = 0; i < ITERS; i++) {
-            // Create database
-            DSLContext conn = DBUtils.getConn();
-            DCMScheduler runner = new DCMScheduler(conn, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-                NUM_APPS, null, false, false);
-
-            // This calls check fill - so some checks just from calling it.
-            runner.fillRandom(FILL_UTIL);
-            assertEquals((runner.coreCapacity() * FILL_UTIL) / 100, runner.usedCores());
-            assertEquals((runner.memsliceCapacity() * FILL_UTIL) / 100, runner.usedMemslices());
-        
-            for (int j = 1; j <= NUM_NODES; j++) {
-                aggregate_core_fill[j - 1] += runner.usedCoresForNode((long) j);
-                aggregate_memslice_fill[j - 1] += runner.usedMemslicesForNode((long) j);
-            }
-
-            for (int j = 1; j <= NUM_APPS; j++) {
-                aggregate_core_per_app[j - 1] += runner.usedCoresForApplication((long) j);
-                aggregate_memslice_per_app[j - 1] += runner.usedMemslicesForApplication((long) j);
-            }
-        }
-
-        float avg_core_fill = 0;
-        float avg_memslice_fill = 0;
-        for (int j = 1; j <= NUM_NODES; j++) {
-            avg_core_fill += (float) aggregate_core_fill[j - 1] / (float) ITERS;
-            avg_memslice_fill += (float) aggregate_memslice_fill[j - 1] / (float) ITERS;
-        }
-        avg_core_fill /= (float) NUM_NODES;
-        avg_memslice_fill /= (float) NUM_NODES;
-
-        // On average, each node should be filled with cores and memslices to FILL_UTIL.
-        // Since it's random, there's a chance this will fail and everything is okay.
-        assertEquals(avg_core_fill, CORES_PER_NODE * ((float) FILL_UTIL / 100.0), 1.0);
-        assertEquals(avg_memslice_fill, MEMSLICES_PER_NODE * ((float) FILL_UTIL / 100.0), 1.0);
-
-        float avg_core_for_app = 0;
-        float avg_memslice_for_app = 0;
-        for (int j = 1; j <= NUM_APPS; j++) {
-            avg_core_for_app += (float) aggregate_core_per_app[j - 1] / (float) ITERS;
-            avg_memslice_for_app += (float) aggregate_memslice_per_app[j - 1] / (float) ITERS;
-        }
-        avg_core_for_app /= (float) NUM_APPS;
-        avg_memslice_for_app /= (float) NUM_APPS;
-
-        // On average, each application should receive (TOTAL_CLUSTER_RESOURCE * FILL%) / NUM_APPS
-        // Since it's random, there's a chance this will fail and everything is okay.
-        assertEquals(avg_core_for_app, 
-                CORES_PER_NODE * NUM_NODES * ((float) FILL_UTIL / 100.0) / (float) NUM_APPS, 1.0);
-        assertEquals(avg_memslice_for_app, 
-                MEMSLICES_PER_NODE * NUM_NODES * ((float) FILL_UTIL / 100.0) / (float) NUM_APPS, 1.0);
-    }
-
-    @Test
-    public void testFillSingleStep() throws Exception {
-        final long NUM_NODES = 4;
-        final long CORES_PER_NODE = 4;
-        final long MEMSLICES_PER_NODE = 4;
-        final long NUM_APPS = 4;
-
-        final long ITERS = 10;
-        final int FILL_UTIL = 50;
-
-        int[] aggregate_core_per_app = new int[(int) NUM_APPS];
-        int[] aggregate_memslice_per_app = new int[(int) NUM_APPS];
-
-        for (int i = 0; i <= ITERS; i++) {
-            // Create database
-            DSLContext conn = DBUtils.getConn();
-            DCMScheduler runner = new DCMScheduler(conn, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-                NUM_APPS, null, false, false);
-
-            runner.fillSingleStep(FILL_UTIL);
-
-            // Make sure filled half way, assume capacities are even
-            assertEquals(runner.coreCapacity() / 2, runner.usedCores());
-            assertEquals(runner.memsliceCapacity() / 2, runner.usedMemslices());
-
-            for (int j = 1; j <= NUM_APPS; j++) {
-                aggregate_core_per_app[j - 1] += runner.usedCoresForApplication((long) j);
-                aggregate_memslice_per_app[j - 1] += runner.usedMemslicesForApplication((long) j);
-            }
-        }
-
-        float avg_core_for_app = 0;
-        float avg_memslice_for_app = 0;
-        for (int j = 1; j <= NUM_APPS; j++) {
-            avg_core_for_app += (float) aggregate_core_per_app[j - 1] / (float) ITERS;
-            avg_memslice_for_app += (float) aggregate_memslice_per_app[j - 1] / (float) ITERS;
-        }
-        avg_core_for_app /= (float) NUM_APPS;
-        avg_memslice_for_app /= (float) NUM_APPS;
-
-        // On average, each application should receive (TOTAL_CLUSTER_RESOURCE * FILL%) / NUM_APPS
-        // Since it's random, there's a chance this will fail and everything is okay.
-        assertEquals(avg_core_for_app, 
-                CORES_PER_NODE * NUM_NODES * ((float) FILL_UTIL / 100.0) / (float) NUM_APPS, 1.0);
-        assertEquals(avg_memslice_for_app, 
-                MEMSLICES_PER_NODE * NUM_NODES * ((float) FILL_UTIL / 100.0) / (float) NUM_APPS, 1.0);
-    }
-
-    @Test
-    public void testFillPoissonUtil() throws Exception {
-        final long NUM_NODES = 10;
-        final long CORES_PER_NODE = 10;
-        final long MEMSLICES_PER_NODE = 20;
-        final long NUM_APPS = 10;
-
-        // Test randomness of fill per node.
-        int aggregate_core_fill = 0;
-        int aggregate_memslice_fill = 0;
-
-        final long ITERS = 10;
-        final int FILL_UTIL = 50;
-
-        // Assume mean requests is for 20% of the cluster per step. This is set high
-        // to reduce the number of steps needed to meet the fill.
-        final double CORE_MEAN = (double) CORES_PER_NODE * 0.2;
-        final double MEMSLICE_MEAN = (double) MEMSLICES_PER_NODE * 0.2;
-
-        int[] aggregate_core_per_app = new int[(int) NUM_APPS];
-        int[] aggregate_memslice_per_app = new int[(int) NUM_APPS];
-
-        for (int i = 0; i < ITERS; i++) {
-            // Create database
-            DSLContext conn = DBUtils.getConn();
-            DCMScheduler runner = new DCMScheduler(conn, NUM_NODES, CORES_PER_NODE, MEMSLICES_PER_NODE, 
-                NUM_APPS, null, false, false);
-
-            // This calls check fill - so some checks just from calling it.
-            runner.fillPoisson(FILL_UTIL, CORE_MEAN, MEMSLICE_MEAN);
-        
-            aggregate_core_fill += runner.usedCores();
-            aggregate_memslice_fill += runner.usedMemslices();
-
-            for (int j = 1; j <= NUM_APPS; j++) {
-                aggregate_core_per_app[j - 1] += runner.usedCoresForApplication((long) j);
-                aggregate_memslice_per_app[j - 1] += runner.usedMemslicesForApplication((long) j);
-            }
-        }
-        float avg_core_fill = ((float) aggregate_core_fill) / ((float) ITERS);
-        float avg_memslice_fill = ((float) aggregate_memslice_fill) / ((float) ITERS);
-
-        // On average, each node should be filled with cores and memslices to FILL_UTIL.
-        // allow for error equal to mean (standard deviation of poisson)
-        assertEquals(avg_core_fill, NUM_NODES * CORES_PER_NODE * ((float) FILL_UTIL / 100.0), CORE_MEAN);
-        assertEquals(avg_memslice_fill, NUM_NODES * MEMSLICES_PER_NODE * ((float) FILL_UTIL / 100.0), MEMSLICE_MEAN);
-
-        float avg_core_for_app = 0;
-        float avg_memslice_for_app = 0;
-        for (int j = 1; j <= NUM_APPS; j++) {
-            avg_core_for_app += (float) aggregate_core_per_app[j - 1] / (float) ITERS;
-            avg_memslice_for_app += (float) aggregate_memslice_per_app[j - 1] / (float) ITERS;
-        }
-        avg_core_for_app /= (float) NUM_APPS;
-        avg_memslice_for_app /= (float) NUM_APPS;
-
-        // On average, each application should receive (TOTAL_CLUSTER_RESOURCE * FILL%) / NUM_APPS
-        // Since it's random, there's a chance this will fail and everything is okay.
-        assertEquals(avg_core_for_app, 
-                CORES_PER_NODE * NUM_NODES * ((float) FILL_UTIL / 100.0) / (float) NUM_APPS, 1.0);
-        assertEquals(avg_memslice_for_app, 
-                MEMSLICES_PER_NODE * NUM_NODES * ((float) FILL_UTIL / 100.0) / (float) NUM_APPS, 1.0);
-    }
-    */
 }
